lib.c: simplify locale handling, and fix some bugs
The previous code incorrectly assumed that setlocale() returned the old
locale; it does not!

Also, it assumed that the call to setlocale() succeeded; it may not!

Only setlocale for category LC_PAPER, and don’t bother resetting on exit.

Fix calls to setlocale() in readspecs() to correctly save and restore the
locale.