More licensing fixes
Change the license of `paper' back to GPL, as it uses GPL-only gnulib
routines. This has to be GPLv3 (so make it GPLv3+), to match gnulib.

Move the relocatable-script module from libgnu to src/libgnu, where it’s
actually used, and license libgnu as LGPLv2.

There’s a remaining problem: paperconfig is GPLv2 only, but the
relocatable-script module is GPLv3+.